2018/prequals/01_More_complicated_than_it_should_have_been/chall/requirements.txt
Django==1.11.15
PyJWT==1.6.4